% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_sentiment.R
\name{article_sentiment}
\alias{article_sentiment}
\title{News Article Sentiment}
\arguments{
\item{url}{Takes character as input. This allows the script to access the website.}
}
\value{
Returns an object of class 'data.table'.
}
\description{
Performs Sentiment Analysis of news articles from the Post-Courier news page
}
\examples{
url <- "https://www.postcourier.com.pg/supplementary-budget-not-a-rushed-secret/";
df <- article_sentiment(url);
print(df)
}
